fix things to make it work for windows linux and mac

+++ core/src/main/java/com/crawljax/browser/WebDriverBackedEmbeddedBrowser.java
@@ -939,10 +939,11 @@ public final class WebDriverBackedEmbeddedBrowser implements EmbeddedBrowser {
             throw new IllegalStateException("CDP is not enabled but CDP screenshot called");
         }
         ChromeDriver chromeDriver = (ChromeDriver) browser;
-        long width = (long) chromeDriver.executeScript("return document.body.scrollWidth");
-        long height = (long) chromeDriver.executeScript("return document.body.scrollHeight");
-        long scale = (long) chromeDriver.executeScript("return window.devicePixelRatio");
-        int intscale = (int) scale;
+        Long width = (Long) chromeDriver.executeScript("return document.body.scrollWidth");
+        Long height = (Long) chromeDriver.executeScript("return document.body.scrollHeight");
+        // Correctly handle devicePixelRatio as Double
+        Double scale = (Double) chromeDriver.executeScript("return window.devicePixelRatio");
+        int intscale = scale.intValue();
 
         HashMap<String, Object> setDeviceMetricsOverride = new HashMap<>();
         setDeviceMetricsOverride.put("deviceScaleFactor", scale);
